third_party/blink/web_tests/external/wpt/mathml/relations/css-styling/lengths-vi-vb-units-expected.txt

This is a testharness.js-based test.
[FAIL] vi/vb in vertical-rl writing modes
  assert_approx_equals: vi in vertical writing modes expected 40 +/- 0.5 but got 64
Harness: the test ran to completion.
